Add RISC-V 'virt' machines support
Add a new JSON file to describe the configurations we want to run.

For RISC-V, the 'virt' machine runs in both qemu-system-riscv32 and
qemu-system-riscv64 binaries, so the "machine" name is working as an
alias (virt32 and virt64) that we need to convert back before running
QEMU.

Disk and network configurations had to be changed to work with these
machines. An "append" option was also needed. This option was proven not
trivial to be moved to the JSON file, probably due to the whitespaces
that the string might contain, that would confuse the parsing done and
mess up the remaining variables.

Signed-off-by: Daniel Henrique Barboza <dbarboza@ventanamicro.com>